Add Prediction Model

+++ LoanPricePrediction/src/main/java/LinearRegressionModel/GetMeanOf_X_Repository.java
@@ -2,8 +2,8 @@ package LinearRegressionModel;
 
 public interface GetMeanOf_X_Repository {
 	
-	public double getMeanOfIncome();
-	public double getMeanOfLoanAmt();
+	public float getMeanOfIncome();
+	public float getMeanOfLoanAmt();
 	public float getMeanOfAge();
 	public float getMeanOfCreditScore();
